extract processor code
