+++ packages/docusaurus-mdx-loader/src/processor.ts
@@ -1,0 +1,223 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import path from 'path';
+import emoji from 'remark-emoji';
+import headings from './remark/headings';
+import toc from './remark/toc';
+import transformImage from './remark/transformImage';
+import transformLinks from './remark/transformLinks';
+import details from './remark/details';
+import head from './remark/head';
+import mermaid from './remark/mermaid';
+import transformAdmonitions from './remark/admonitions';
+import codeCompatPlugin from './remark/mdx1Compat/codeCompatPlugin';
+import type {MDXFrontMatter} from './frontMatter';
+import type {Options} from './loader';
+
+// @ts-expect-error: TODO see https://github.com/microsoft/TypeScript/issues/49721
+import type {Processor} from 'unified';
+import type {AdmonitionOptions} from './remark/admonitions';
+
+// @ts-expect-error: TODO see https://github.com/microsoft/TypeScript/issues/49721
+import type {ProcessorOptions} from '@mdx-js/mdx';
+
+// TODO as of April 2023, no way to import/re-export this ESM type easily :/
+// This might change soon, likely after TS 5.2
+// See https://github.com/microsoft/TypeScript/issues/49721#issuecomment-1517839391
+type Pluggable = any; // TODO fix this asap
+
+// Copied from https://mdxjs.com/packages/mdx/#optionsmdextensions
+// Although we are likely to only use .md / .mdx anyway...
+const mdFormatExtensions = [
+  '.md',
+  '.markdown',
+  '.mdown',
+  '.mkdn',
+  '.mkd',
+  '.mdwn',
+  '.mkdown',
+  '.ron',
+];
+
+function isMDFormat(filepath: string) {
+  return mdFormatExtensions.includes(path.extname(filepath));
+}
+
+function getFormat({
+  filePath,
+  frontMatterFormat,
+}: {
+  filePath: string;
+  frontMatterFormat: MDXFrontMatter['format'];
+}): 'md' | 'mdx' {
+  return frontMatterFormat === 'detect'
+    ? isMDFormat(filePath)
+      ? 'md'
+      : 'mdx'
+    : frontMatterFormat;
+}
+
+const DEFAULT_OPTIONS: MDXOptions = {
+  admonitions: true,
+  rehypePlugins: [],
+  remarkPlugins: [emoji, headings, toc],
+  beforeDefaultRemarkPlugins: [],
+  beforeDefaultRehypePlugins: [],
+};
+
+// We use different compilers depending on the file type (md vs mdx)
+type Processors = {
+  mdCompiler: Processor;
+  mdxCompiler: Processor;
+};
+
+// Compilers are cached so that Remark/Rehype plugins can run
+// expensive code during initialization
+const compilersCache = new Map<string | Options, Processors>();
+
+export type MDXPlugin = Pluggable;
+
+export type MDXOptions = {
+  admonitions: boolean | Partial<AdmonitionOptions>;
+  remarkPlugins: MDXPlugin[];
+  rehypePlugins: MDXPlugin[];
+  beforeDefaultRemarkPlugins: MDXPlugin[];
+  beforeDefaultRehypePlugins: MDXPlugin[];
+};
+
+function getAdmonitionsPlugins(
+  admonitionsOption: MDXOptions['admonitions'],
+): MDXPlugin[] {
+  if (admonitionsOption) {
+    const plugin: MDXPlugin =
+      admonitionsOption === true
+        ? transformAdmonitions
+        : [transformAdmonitions, admonitionsOption];
+    return [plugin];
+  }
+
+  return [];
+}
+
+export async function getProcessors({
+  query,
+  reqOptions,
+}: {
+  query: string | Options;
+  reqOptions: Options;
+}): Promise<Processors> {
+  const {createProcessor} = await import('@mdx-js/mdx');
+  const {default: rehypeRaw} = await import('rehype-raw');
+  const {default: gfm} = await import('remark-gfm');
+  const {default: comment} = await import('remark-comment');
+  const {default: directives} = await import('remark-directive');
+
+  const compilers = compilersCache.get(query);
+  if (compilers) {
+    return compilers;
+  }
+
+  // /!\ this method is synchronous on purpose
+  // Using async code here can create cache entry race conditions!
+  function createCompilersSynchronous(): Processors {
+    const remarkPlugins: MDXPlugin[] = [
+      ...(reqOptions.beforeDefaultRemarkPlugins ?? []),
+      directives,
+      ...getAdmonitionsPlugins(reqOptions.admonitions ?? false),
+      ...DEFAULT_OPTIONS.remarkPlugins,
+      details,
+      head,
+      ...(reqOptions.markdownConfig.mermaid ? [mermaid] : []),
+      [
+        transformImage,
+        {
+          staticDirs: reqOptions.staticDirs,
+          siteDir: reqOptions.siteDir,
+        },
+      ],
+      [
+        transformLinks,
+        {
+          staticDirs: reqOptions.staticDirs,
+          siteDir: reqOptions.siteDir,
+        },
+      ],
+      gfm,
+      reqOptions.markdownConfig.mdx1Compat.comments ? comment : null,
+      ...(reqOptions.remarkPlugins ?? []),
+    ].filter((plugin): plugin is MDXPlugin => Boolean(plugin));
+
+    // codeCompatPlugin needs to be applied last after user-provided plugins
+    // (after npm2yarn for example)
+    remarkPlugins.push(codeCompatPlugin);
+
+    // This is what permits to embed HTML elements with format 'md'
+    // See https://github.com/mdx-js/mdx/pull/2295#issuecomment-1540085960
+    const rehypeRawPlugin: MDXPlugin = [
+      rehypeRaw,
+      {
+        passThrough: [
+          'mdxFlowExpression',
+          'mdxJsxFlowElement',
+          'mdxJsxTextElement',
+          'mdxTextExpression',
+          'mdxjsEsm',
+        ],
+      },
+    ];
+
+    const rehypePlugins: MDXPlugin[] = [
+      rehypeRawPlugin,
+      ...(reqOptions.beforeDefaultRehypePlugins ?? []),
+      ...DEFAULT_OPTIONS.rehypePlugins,
+      ...(reqOptions.rehypePlugins ?? []),
+    ];
+
+    const options: ProcessorOptions & Options = {
+      ...reqOptions,
+      remarkPlugins,
+      rehypePlugins,
+      providerImportSource: '@mdx-js/react',
+    };
+
+    return {
+      mdCompiler: createProcessor({
+        ...options,
+        rehypePlugins: [rehypeRawPlugin, ...(options.rehypePlugins ?? [])],
+        format: 'md',
+      }),
+      mdxCompiler: createProcessor({
+        ...options,
+        format: 'mdx',
+      }),
+    };
+  }
+
+  const compilerCacheEntry = createCompilersSynchronous();
+  compilersCache.set(query, compilerCacheEntry);
+  return compilerCacheEntry;
+}
+
+export async function getProcessor({
+  filePath,
+  mdxFrontMatter,
+  query,
+  reqOptions,
+}: {
+  filePath: string;
+  mdxFrontMatter: MDXFrontMatter;
+  query: string | Options;
+  reqOptions: Options;
+}): Promise<Processor> {
+  const compilers = await getProcessors({query, reqOptions});
+  const format = getFormat({
+    filePath,
+    frontMatterFormat: mdxFrontMatter.format,
+  });
+  return format === 'md' ? compilers.mdCompiler : compilers.mdxCompiler;
+}
